DF-6961 exitCode long
Former-commit-id: c52f13c7071f0ea11d3067b9e0050677fc999864

+++ Sources/Internal/Platform/Process.h
@@ -81,7 +81,7 @@ private:
     int pipes[2];
 #endif
 
-    int exitCode;
+    long exitCode;
     
 };
 };
